move each filter to a seperate directory
