+++ packages/lib/src/spam-filter/filter/nonceFilter/NonceFilter.test.ts
@@ -1,4 +1,4 @@
-import { DeliveryInformation } from '../../messaging';
+import { DeliveryInformation } from '../../../messaging';
 import { nonceFilter } from './NonceFilter';
 
 test('Should accept a message with a nonce equal the nonce threshold', async () => {
